% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{find_authors}
\alias{find_authors}
\title{Extract author information}
\usage{
find_authors(file_path)
}
\arguments{
\item{file_path}{A file from JSTOR containing meta-data which can be read in with
\code{\link[xml2:read_xml]{xml2::read_xml()}}. \code{find_authors} takes care of importing the file.}
}
\value{
A \code{tibble} containing the extracted authors
}
\description{
\code{find_authors()} extracts information about authors from JSTOR-XML files.
}
\details{
\strong{Authors}:

A list-column with four possible columns:
\itemize{
\item given_name
\item surname
\item author_number among co-authors
\item string_name In some cases the name is not availible in separate fields, then
the full name is just available as a full string.
}

With this approach some information is omitted: some authors may have a suffix
to their name, like \emph{Jr}, or like \emph{III} (i.e. William the Third).
Those are currently not being extracted.
}
\examples{
find_authors(jstor_example("sample_with_references.xml"))
}
