% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R
\name{find_footnotes}
\alias{find_footnotes}
\title{Extract all footnotes}
\usage{
find_footnotes(file_path)
}
\arguments{
\item{file_path}{The path to the \code{.xml}-file from which footnotes should be
extracted.}
}
\value{
A \code{tibble} containing the content from \code{fn-group} (usually the
footnotes). If there were no footnotes, \code{NA_character} is returned for the
column \code{footnotes}.
}
\description{
This function extracts the content of \code{fn-group} from journal-articles.
}
\examples{
find_footnotes(jstor_example("sample_with_footnotes.xml"))
}
