% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{find_meta}
\alias{find_meta}
\title{Extract meta information}
\usage{
find_meta(file_path)
}
\arguments{
\item{file_path}{A file from JSTOR containing meta-data which can be read in with
\code{\link[xml2:read_xml]{xml2::read_xml()}}. \code{find_meta} takes care of importing the file.}
}
\value{
A \code{tibble} containing the extracted meta-data.
}
\description{
\code{find_meta()} extracts meta-data from JSTOR-XML files.
}
\details{
\strong{Publication dates}:

Always the first entry is being extracted, which should correspond to the oldest
date, in case there is more than one date.
}
\examples{
find_meta(jstor_example("sample_with_references.xml"))
}
