% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentations.R
\name{jst_add_total_pages}
\alias{jst_add_total_pages}
\title{Add total count of pages}
\usage{
jst_add_total_pages(meta_data, quietly = FALSE)
}
\arguments{
\item{meta_data}{Data which was processed via \code{\link[=jst_get_article]{jst_get_article()}}.}

\item{quietly}{Should warnings from converting page ranges be suppressed?}
}
\value{
A \code{tibble}, as provided with in \code{meta_data}, with an additional
column on total number of pages.
}
\description{
This function adds a column with the total count of pages. It calls
\code{\link[=jst_get_total_pages]{jst_get_total_pages()}} which does the main work.
}
\seealso{
\code{\link[=jst_get_total_pages]{jst_get_total_pages()}}
}
