% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{jst_get_file_name}
\alias{jst_get_file_name}
\title{Extract the basename of a path}
\usage{
jst_get_file_name(file_path)
}
\arguments{
\item{file_path}{A path to a file}
}
\description{
This helper simply extracts the basename of a path and removes the extension,
e.g. \code{foo/bar.txt} is shortened to \code{bar}.
}
