% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_text.R
\name{jst_read_fulltext}
\alias{jst_read_fulltext}
\title{Import full-text}
\usage{
jst_read_fulltext(filename)
}
\arguments{
\item{filename}{The path to the file.}
}
\value{
A \code{tibble}, containing the file-path as id, the full content of
the file, and the encoding which was used to read it.
}
\description{
This function imports the full_text contents of a JSTOR-article.
}
