% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstor_package.R
\docType{package}
\name{jstor}
\alias{jstor}
\title{jstor: Read Data from JSTOR/DfR}
\description{
The tool \href{https://www.jstor.org/dfr/}{Data for Research (DfR)} by JSTOR is a
valuable source for citation analysis and text mining. \code{jstor}
provides functions and suggests workflows for importing
datasets from DfR.
}
\details{
Please refer to the vignettes for information on how to use the package:

\code{browseVignettes("jstor")}

If you encounter any issues or have ideas for new features, please file an
issue at \url{https://github.com/ropensci/jstor/issues}.
}
\author{
Thomas Klebel
}
