% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_import_fun.R
\name{jstor_convert_to_file}
\alias{jstor_convert_to_file}
\title{Batch import data}
\usage{
jstor_convert_to_file(file_paths, chunk_number, path, fun, col_names = FALSE,
  cores = getOption("mc.cores", 1L))
}
\arguments{
\item{file_paths}{A character vector of file-paths.}

\item{chunk_number}{An integer, specifying the number of the chunk. Will be
appended to the output-file.}

\item{path}{The path where the files should be written to. Should include the
basename for the files too, i.e. \code{path/to/outfiles/meta_data}.}

\item{fun}{The function to use when importing files, i.e. \code{find_meta},
\code{find_authors} and \code{read_full_text}.}

\item{col_names}{Should \code{col_names} be printed when exporting results? For
errors, col_names are always written to file.}

\item{cores}{Number of cores to use for parallel processing.}
}
\description{
This function does the heavy lifting of importing all data.
}
