% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{jstor_example}
\alias{jstor_example}
\title{Defunct: Get path to jstor example}
\usage{
jstor_example(path = NULL)
}
\arguments{
\item{path}{Name of the example file. If \code{NULL}, the example files will be
listed.}
}
\description{
This function will be removed in the next release. Please use
\code{jst_example()} instead.
}
\details{
jstor includes several sample files for demonstration purposes. This helper
makes them easy to access.

The code for this function was adapted from the package \code{readr}.
}
\examples{
\dontrun{
jstor_example()
jstor_example("sample_with_references.xml") 
}
}
