% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_import_fun.R
\name{jstor_import}
\alias{jstor_import}
\title{Wrapper for file import}
\usage{
jstor_import(in_paths, out_file, out_path = NULL, .f, col_names = TRUE,
  files_per_batch = 10000, cores = getOption("mc.cores", 1L),
  show_progress = TRUE)
}
\arguments{
\item{in_paths}{A character vector to the \code{xml}-files which should be
imported}

\item{out_file}{Name of files to export to. Each batch gets appended by an
increasing number.}

\item{out_path}{Path to export files to (combined with filename).}

\item{.f}{Function to use for import. Can be one of \code{find_article},
\code{find_authors}, \code{find_references}, \code{find_footnotes}, \code{find_book} or
\code{find_chapter}.}

\item{col_names}{Should column names be written to file? Defaults to \code{TRUE}.}

\item{files_per_batch}{Number of files for each batch.}

\item{cores}{Number of cores to use for parallel processing.}

\item{show_progress}{Displays a progress bar for each batch, if the session
is interactive.}
}
\value{
Writes \code{.csv}-files to disk.
}
\description{
This function applies an import function to a list of \code{xml}-files and saves
them in batches of \code{.csv}-files to disk.
}
