% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_import_fun.R
\name{jstor_import_wrapper}
\alias{jstor_import_wrapper}
\title{Wrapper for file conversion}
\usage{
jstor_import_wrapper(file_paths, out_file, path = NULL, .f,
  files_per_batch = 4000, cores = getOption("mc.cores", 2L))
}
\arguments{
\item{file_paths}{A character vector to the \code{xml}-files which should be imported}

\item{out_file}{Name of files to export to. Each batch gets appended by an
increasing number.}

\item{path}{Path to export files to (combined with filename).}

\item{.f}{Function to use for import. Can be one of \code{find_meta}, \code{find_authors},
\code{find_references} or \code{find_footnotes}.}

\item{files_per_batch}{Number of files for each batch.}

\item{cores}{Number of cores to use for parallel processing.}
}
\value{
Writes .csv-files to disk.
}
\description{
This function applies an import function to a list of \code{xml}-files and saves
them in batches of .csv-files to disk.
}
