% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.jstor}
\alias{print.jstor}
\title{Print results in a pretty way}
\usage{
\method{print}{jstor}(x, ...)
}
\arguments{
\item{x}{A jstor object, created via one of the \code{find_*}-functions.}
}
\description{
Print results in a pretty way
}
\keyword{internal}
