% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{read_full_text}
\alias{read_full_text}
\title{Defunct: Import full-text}
\usage{
read_full_text(filename)
}
\arguments{
\item{filename}{The path to the file.}
}
\value{
A \code{tibble}, containing the file-path as id, the full content of
the file, and the encoding which was used to read it.
}
\description{
This function will be removed in the next release. Please use
\code{jst_get_full_text()} instead.
}
\details{
This function imports the full_text contents of a JSTOR-article.
}
