% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R
\name{find_footnotes}
\alias{find_footnotes}
\title{Extract all footnotes}
\usage{
find_footnotes(file_path)
}
\arguments{
\item{file_path}{The path to the xml-file from which footnotes should be
extracted.}
}
\value{
A \code{tibble} containing the footnotes
}
\description{
This function extracts the content of \code{fn-group} within the \code{xml}-file. If
there are no footnotes, \code{NA} is returend.
}
