% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentations.R
\name{jst_get_total_pages}
\alias{jst_get_total_pages}
\title{Find total pages}
\usage{
jst_get_total_pages(first_page = NULL, last_page = NULL,
  page_range = NULL)
}
\arguments{
\item{first_page}{The first page of an article (numeric).}

\item{last_page}{The last page of an article (numeric).}

\item{page_range}{The page range of an article (character).}
}
\value{
A vector with the calculated total pages.
}
\description{
This function is a simple helper to calculate the total number of pages of
an article.
}
\details{
This function deals with three cases:
\itemize{
\item if page_range is supplied, the number of pages is calculated from it.
\item if only the first page is supplied, NA is returned
\item if first and last page are supplied, the number of pages is calculated as
\code{last_page - first_page + 1}
}
}
\examples{
# calculate pages from first and last page
first_pages <- sample(30:50, 10)
last_pages <- first_pages + sample(5:20, 10)

jst_get_total_pages(first_pages, last_pages)

# get pages from page range
jst_get_total_pages(page_range = "51 - 70, 350")
}
