% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{jstor_example}
\alias{jstor_example}
\title{Get path to jstor example}
\usage{
jstor_example(path = NULL)
}
\arguments{
\item{path}{Name of the example file. If \code{NULL}, the example files will be
listed.}
}
\description{
jstor includes several sample files for demonstration purposes. This helper
makes them easy to access.
}
\details{
The code for this function was adapted from the package \code{readr}.
}
\examples{
jstor_example()
jstor_example("sample_with_references.xml") 
}
